# Conflicts:
#	domain_config_ui/src/Form/SwitchForm.php
#	domain_source/domain_source.module
#	domain_source/src/DomainSourceElementManager.php

+++ domain/src/Commands/DomainCommands.php
@@ -6,13 +6,12 @@ use Consolidation\AnnotatedCommand\Events\CustomEventAwareInterface;
 use Consolidation\AnnotatedCommand\Events\CustomEventAwareTrait;
 use Consolidation\OutputFormatters\StructuredData\PropertyList;
 use Consolidation\OutputFormatters\StructuredData\RowsOfFields;
-use Drupal\Core\Config\StorageException;
-use Drupal\Core\Entity\EntityStorageException;
-use Drupal\Core\Entity\FieldableEntityInterface;
 use Drupal\Component\Plugin\Exception\InvalidPluginDefinitionException;
 use Drupal\Component\Plugin\Exception\PluginException;
 use Drupal\Component\Plugin\Exception\PluginNotFoundException;
 use Drupal\Component\Utility\Html;
+use Drupal\Core\Entity\EntityStorageException;
+use Drupal\Core\Entity\FieldableEntityInterface;
 use Drupal\domain\DomainInterface;
 use Drush\Commands\DrushCommands;
 use GuzzleHttp\Exception\TransferException;
@@ -93,7 +92,7 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
     // Load all domains:
     $domains = $this->domainStorage()->loadMultipleSorted();
 
-    if (empty($domains)) {
+    if (count($domains) === 0) {
       $this->logger()->warning(dt('No domains have been created. Use "drush domain:add" to create one.'));
       return new RowsOfFields([]);
     }
@@ -134,16 +133,16 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
             break;
 
           case 'status':
-            $v = $domain->get($key);
-            if (($options['inactive'] && $v) || ($options['active'] && !$v)) {
+            $value = (bool) $domain->get($key);
+            if (($options['inactive'] && $value) || ($options['active'] && !$value)) {
               continue 3;
             }
-            $v = !empty($v) ? dt('Active') : dt('Inactive');
+            $v = $value ? dt('Active') : dt('Inactive');
             break;
 
           case 'is_default':
-            $v = $domain->get($key);
-            $v = !empty($v) ? dt('Default') : '';
+            $value = (bool) $domain->get($key);
+            $v = $value ? dt('Default') : '';
             break;
 
           default:
@@ -184,7 +183,8 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
   public function infoDomains() {
     $default_domain = $this->domainStorage()->loadDefaultDomain();
 
-    // Load all domains:
+    // Load all domains.
+    /** @var \Drupal\domain\DomainInterface[] $all_domains */
     $all_domains = $this->domainStorage()->loadMultiple(NULL);
     $active_domains = [];
     foreach ($all_domains as $domain) {
@@ -214,14 +214,14 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
 
         case 'default_id':
           $v = '-unset-';
-          if ($default_domain) {
+          if ($default_domain instanceof DomainInterface) {
             $v = $default_domain->id();
           }
           break;
 
         case 'default_host':
           $v = '-unset-';
-          if ($default_domain) {
+          if ($default_domain instanceof DomainInterface) {
             $v = $default_domain->getHostname();
           }
           break;
@@ -297,25 +297,29 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
    *
    * @throws \Drupal\domain\Commands\DomainCommandException
    */
-  public function add($hostname, $name, array $options = [
-    'weight' => NULL,
-    'scheme' => NULL
-  ]) {
+  public function add(
+    $hostname,
+    $name,
+    array $options = [
+      'weight' => NULL,
+      'scheme' => NULL,
+    ],
+  ) {
     // Validate the weight arg.
-    if (!empty($options['weight']) && !is_numeric($options['weight'])) {
+    if (!is_null($options['weight']) && !is_numeric($options['weight'])) {
       throw new DomainCommandException(
         dt('Domain weight "!weight" must be a number',
-          ['!weight' => !empty($options['weight']) ? $options['weight'] : ''])
+          ['!weight' => $options['weight'] ?? 'null'])
       );
     }
 
     // Validate the scheme arg.
-    if (!empty($options['scheme']) &&
+    if (!is_null($options['scheme']) &&
       ($options['scheme'] !== 'http' && $options['scheme'] !== 'https' && $options['scheme'] !== 'variable')
     ) {
       throw new DomainCommandException(
         dt('Scheme name "!scheme" not known',
-          ['!scheme' => !empty($options['scheme']) ? $options['scheme'] : ''])
+          ['!scheme' => $options['scheme'] ?? 'null'])
       );
     }
 
@@ -324,10 +328,10 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
     $values = [
       'hostname' => $hostname,
       'name' => $name,
-      'status' => empty($options['inactive']),
-      'scheme' => empty($options['scheme']) ? 'http' : $options['scheme'],
-      'weight' => empty($options['weight']) ? $start_weight : $options['weight'],
-      'is_default' => !empty($options['is_default']),
+      'status' => $options['inactive'] ?? 1,
+      'scheme' => $options['scheme'] ?? 'http',
+      'weight' => $options['weight'] ?? $start_weight,
+      'is_default' => $options['is_default'] ?? 0,
       'id' => $this->domainStorage()->createMachineName($hostname),
     ];
     /** @var \Drupal\domain\DomainInterface */
@@ -335,7 +339,7 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
 
     // Check for hostname validity. This is required.
     $valid = $this->validateDomain($domain);
-    if (!empty($valid)) {
+    if (count($valid) > 0) {
       throw new DomainCommandException(
         dt('Hostname is not valid. !errors',
           ['!errors' => implode(" ", $valid)])
@@ -430,11 +434,14 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
    *
    * @see https://github.com/consolidation/annotated-command#option-event-hook
    */
-  public function delete($domain_id, array $options = [
-    'users-assign' => NULL,
-    'dryrun' => NULL,
-    'chatty' => NULL
-  ]) {
+  public function delete(
+    $domain_id,
+    array $options = [
+      'users-assign' => NULL,
+      'dryrun' => NULL,
+      'chatty' => NULL,
+    ],
+  ) {
 
     $message = '';
     $messages = [];
@@ -445,10 +452,10 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
     // Get current domain list and perform validation checks.
     $all_domains = $this->domainStorage()->loadMultipleSorted();
 
-    if (empty($all_domains)) {
+    if (count($all_domains) === 0) {
       throw new DomainCommandException('There are no configured domains.');
     }
-    if (empty($domain_id)) {
+    if ($domain_id === '') {
       throw new DomainCommandException('You must specify a domain to delete.');
     }
 
@@ -456,25 +463,28 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
     if ($domain_id === 'all') {
       $domains = $all_domains;
       $really = $this->io()->confirm(dt('This action cannot be undone. Continue?:'), FALSE);
-      if (empty($really)) {
+      if (!$really) {
         return;
       }
       $message = dt('All domain records have been deleted.');
     }
-    elseif ($domain = $this->getDomainFromArgument($domain_id)) {
-      if ($domain->isDefault()) {
-        throw new DomainCommandException('The primary domain may not be deleted.
+    else {
+      $domain = $this->getDomainFromArgument($domain_id);
+      if ($domain instanceof DomainInterface) {
+        if ($domain->isDefault()) {
+          throw new DomainCommandException('The primary domain may not be deleted.
           Use drush domain:default to set a new default domain.');
+        }
+        $domains = [$domain];
+        $message = dt('Domain record !domain deleted.',
+          ['!domain' => $domain->id()]
+        );
       }
-      $domains = [$domain];
-      $message = dt('Domain record !domain deleted.',
-        ['!domain' => $domain->id()]
-      );
     }
 
     // Set the reassignment policy.
     $policy_users = 'prompt';
-    if (!empty($options['users-assign'])) {
+    if (isset($options['users-assign']) && $options['users-assign']) {
       if (in_array($options['users-assign'], $this->reassignmentPolicies, TRUE)) {
         $policy_users = $options['users-assign'];
       }
@@ -494,10 +504,11 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
 
     $this->deleteDomain($domains, $options);
 
-    if ($messages) {
+    if (count($messages) > 0) {
       $message .= "\n" . implode("\n", $messages);
     }
     $this->logger()->info($message);
+
     return $message;
   }
 
@@ -537,7 +548,7 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
       $policy = $this->io()->choice(dt('Reassign @type field @field data to:',
         [
           '@type' => $delete_options['entity_filter'],
-          '@field' => $delete_options['field']
+          '@field' => $delete_options['field'],
         ]), $reassign_list);
     }
     elseif ($policy === 'default') {
@@ -587,7 +598,8 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
       $domains = $this->domainStorage()->loadMultipleSorted();
     }
     else {
-      if ($domain = $this->getDomainFromArgument($domain_id)) {
+      $domain = $this->getDomainFromArgument($domain_id);
+      if ($domain instanceof DomainInterface) {
         $domains = [$domain];
       }
       else {
@@ -631,11 +643,15 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
    *
    * @throws \Drupal\domain\Commands\DomainCommandException
    */
-  public function defaultDomain($domain_id, array $options = [
-    'validate' => NULL
-  ]) {
+  public function defaultDomain(
+    $domain_id,
+    array $options = [
+      'validate' => NULL,
+    ],
+  ) {
     // Resolve the domain.
-    if (!empty($domain_id) && $domain = $this->getDomainFromArgument($domain_id)) {
+    $domain = $this->getDomainFromArgument($domain_id);
+    if ($domain instanceof DomainInterface) {
       $validate = ($options['validate']) ? 1 : 0;
       $domain->addProperty('validate_url', $validate);
       if ($error = $this->checkHttpResponse($domain)) {
@@ -679,7 +695,8 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
    */
   public function disable($domain_id) {
     // Resolve the domain.
-    if ($domain = $this->getDomainFromArgument($domain_id)) {
+    $domain = $this->getDomainFromArgument($domain_id);
+    if ($domain instanceof DomainInterface) {
       if ($domain->status()) {
         $domain->disable();
         $this->logger()->info(dt('!domain has been disabled.',
@@ -716,7 +733,8 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
    */
   public function enable($domain_id) {
     // Resolve the domain.
-    if ($domain = $this->getDomainFromArgument($domain_id)) {
+    $domain = $this->getDomainFromArgument($domain_id);
+    if ($domain instanceof DomainInterface) {
       if (!$domain->status()) {
         $domain->enable();
         $this->logger()->info(dt('!domain has been enabled.',
@@ -755,7 +773,8 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
    */
   public function renameDomain($domain_id, $name) {
     // Resolve the domain.
-    if ($domain = $this->getDomainFromArgument($domain_id)) {
+    $domain = $this->getDomainFromArgument($domain_id);
+    if ($domain instanceof DomainInterface) {
       $domain->saveProperty('name', $name);
       return dt('Renamed !domain to !name.',
         ['!domain' => $domain->getHostname(), '!name' => $domain->label()]);
@@ -786,8 +805,9 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
     $new_scheme = NULL;
 
     // Resolve the domain.
-    if ($domain = $this->getDomainFromArgument($domain_id)) {
-      if (!empty($scheme)) {
+    $domain = $this->getDomainFromArgument($domain_id);
+    if ($domain instanceof DomainInterface) {
+      if (!is_null($scheme)) {
         // Set with a value.
         $new_scheme = $scheme;
       }
@@ -802,7 +822,7 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
       }
 
       // If we were asked to change scheme, validate the value and do so.
-      if (!empty($new_scheme)) {
+      if (!is_null($new_scheme)) {
         switch ($new_scheme) {
           case 'http':
             $new_scheme = 'http';
@@ -869,28 +889,30 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
    *
    * @throws \Drupal\domain\Commands\DomainCommandException
    */
-  public function generate($primary = 'example.com', array $options = [
-    'count' => NULL,
-    'empty' => NULL,
-    'scheme' => 'http'
-  ]) {
+  public function generate(
+    $primary = 'example.com',
+    array $options = [
+      'count' => NULL,
+      'empty' => NULL,
+      'scheme' => 'http',
+    ],
+  ) {
     // Check the number of domains to create.
     $count = $options['count'];
     if (is_null($count)) {
       $count = 15;
     }
 
-    $domains = $this->domainStorage()->loadMultiple(NULL);
-    if (!empty($options['empty'])) {
+    $domains = $this->domainStorage()->loadMultiple();
+    if (isset($options['empty']) && $options['empty']) {
       $this->domainStorage()->delete($domains);
-      $domains = $this->domainStorage()->loadMultiple(NULL);
+      $domains = $this->domainStorage()->loadMultiple();
     }
     // Ensure we don't duplicate any domains.
     $existing = [];
-    if (!empty($domains)) {
-      foreach ($domains as $domain) {
-        $existing[] = $domain->getHostname();
-      }
+    /** @var \Drupal\domain\DomainInterface[] $domains */
+    foreach ($domains as $domain) {
+      $existing[] = $domain->getHostname();
     }
     // Set up one.* and so on.
     $names = [
@@ -964,7 +986,7 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
     }
 
     // If nothing created, say so.
-    if (empty($prepared)) {
+    if (count($prepared) === 0) {
       return dt('No new domains were created.');
     }
     else {
@@ -1017,7 +1039,7 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
 
     // Try loading domain assuming arg is a machine name.
     $domain = $this->domainStorage()->load($argument);
-    if (!$domain) {
+    if (is_null($domain)) {
       // Try loading assuming it is a host name.
       $domain = $this->domainStorage()->loadByHostname($argument);
     }
@@ -1025,7 +1047,7 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
     // domain_id (an INT) is only used internally because the Node Access
     // system demands the use of numeric keys. It should never be used to load
     // or identify domain records. Use the machine_name or hostname instead.
-    if (!$domain) {
+    if (is_null($domain)) {
       throw new DomainCommandException(
         dt('Domain record could not be found from "!a".', ['!a' => $argument])
       );
@@ -1050,6 +1072,7 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
   protected function filterDomains(array $domains, array $exclude, array $initial = []) {
     foreach ($domains as $domain) {
       // Exclude unwanted domains.
+      // @phpstan-ignore-next-line
       if (!in_array($domain->id(), $exclude, FALSE)) {
         $initial[$domain->id()] = $domain;
       }
@@ -1112,7 +1135,7 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
         throw new DomainCommandException('Unable to save domain', $e);
       }
 
-      if ($domain->getDomainId()) {
+      if ($domain->getDomainId() > 0) {
         $this->logger()->info(dt('Created @name at @domain.',
           ['@name' => $domain->label(), '@domain' => $domain->getHostname()]));
         return TRUE;
@@ -1180,7 +1203,7 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
       }
 
       try {
-        // Fire any registered hooks for deletion, passing them current imput.
+        // Fire any registered hooks for deletion, passing them current input.
         $handlers = $this->getCustomEventHandlers('domain-delete');
         $messages = [];
         foreach ($handlers as $handler) {
@@ -1246,7 +1269,7 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
    */
   protected function ensureEntityFieldMap() {
     // Try to avoid repeated calls to getFieldMap() assuming it's expensive.
-    if (empty($this->entityFieldMap)) {
+    if (is_null($this->entityFieldMap)) {
       $entity_field_manager = \Drupal::service('entity_field.manager');
       $this->entityFieldMap = $entity_field_manager->getFieldMap();
     }
@@ -1286,6 +1309,7 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
     if ($just_count) {
       $efq->count();
     }
+
     return $efq->execute();
   }
 
@@ -1404,22 +1428,22 @@ class DomainCommands extends DrushCommands implements CustomEventAwareInterface 
   protected function reassignLinkedEntities(array $domains, array $options) {
     $count = 0;
     $field = $options['field'];
-    $entity_typenames = $this->findDomainEnabledEntities($field);
+    $entity_types = $this->findDomainEnabledEntities($field);
 
     $new_domain = $this->getDomainInstanceFromPolicy($options['policy']);
-    if (empty($new_domain)) {
+    if (is_null($new_domain)) {
       throw new DomainCommandException('invalid destination domain');
     }
 
     // Loop through each entity type.
     $exceptions = FALSE;
-    foreach ($entity_typenames as $name) {
-      if (empty($options['entity_filter']) || $options['entity_filter'] === $name) {
+    foreach ($entity_types as $name) {
+      if (!isset($options['entity_filter']) || $options['entity_filter'] === $name) {
 
         // For each domain being reassigned from...
         foreach ($domains as $domain) {
           $ids = $this->enumerateDomainEntities($name, $domain->id(), $field);
-          if (!empty($ids)) {
+          if ($ids !== []) {
             try {
               if ($options['chatty']) {
                 $this->logger()->info('Reassigning @count @entity_name entities to @domain',
